backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
Flask-Bcrypt==1.0.1
Flask-Mail==0.10.0
flask-swagger-ui==4.11.1
python-dotenv==1.0.1
